requirements.txt
streamlit
langchain
databricks-langchain
faiss-cpu
numpy